+++ app/src/androidTest/java/org/mozilla/focus/screengrab/BrowserFragmentScreenshot.kt
@@ -26,8 +26,8 @@ import org.mozilla.focus.autobot.indexOfType
 import org.mozilla.focus.autobot.session
 import org.mozilla.focus.helper.BeforeTestTask
 import org.mozilla.focus.utils.AndroidTestUtils
-import org.mozilla.rocket.content.BottomBarItemAdapter.Companion.TYPE_TAB_COUNTER
-import org.mozilla.rocket.content.BottomBarViewModel.Companion.DEFAULT_BOTTOM_BAR_ITEMS
+import org.mozilla.rocket.chrome.BottomBarItemAdapter.Companion.TYPE_TAB_COUNTER
+import org.mozilla.rocket.chrome.BottomBarViewModel.Companion.DEFAULT_BOTTOM_BAR_ITEMS
 import tools.fastlane.screengrab.FalconScreenshotStrategy
 import tools.fastlane.screengrab.Screengrab
 import java.io.IOException
