+++ app/src/androidTest/java/org/mozilla/focus/utils/AndroidTestUtils.java
@@ -27,7 +27,7 @@ import org.mozilla.focus.activity.MainActivity;
 import org.mozilla.focus.autobot.BottomBarRobot;
 import org.mozilla.focus.autobot.BottomBarRobotKt;
 import org.mozilla.focus.helper.BeforeTestTask;
-import org.mozilla.rocket.content.BottomBarViewModel;
+import org.mozilla.rocket.chrome.BottomBarViewModel;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -49,7 +49,7 @@ import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.core.AllOf.allOf;
 import static org.mozilla.focus.utils.RecyclerViewTestUtils.clickChildViewWithId;
-import static org.mozilla.rocket.content.BottomBarItemAdapter.TYPE_MENU;
+import static org.mozilla.rocket.chrome.BottomBarItemAdapter.TYPE_MENU;
 
 public final class AndroidTestUtils {
