+++ app/src/main/java/org/mozilla/focus/Inject.java
@@ -21,12 +21,7 @@ import org.mozilla.focus.persistence.TabsDatabase;
 import org.mozilla.focus.utils.AppConstants;
 import org.mozilla.focus.utils.RemoteConfigConstants;
 import org.mozilla.focus.utils.Settings;
-import org.mozilla.rocket.content.BottomBarViewModel;
-import org.mozilla.rocket.content.BottomBarViewModelFactory;
-import org.mozilla.rocket.content.ChromeViewModel;
-import org.mozilla.rocket.content.ChromeViewModelFactory;
-import org.mozilla.rocket.content.MenuViewModel;
-import org.mozilla.rocket.content.MenuViewModelFactory;
+import org.mozilla.rocket.chrome.*;
 import org.mozilla.rocket.download.DownloadIndicatorViewModel;
 import org.mozilla.rocket.download.DownloadInfoRepository;
 import org.mozilla.rocket.download.DownloadInfoViewModel;
