+++ app/src/main/java/org/mozilla/focus/activity/MainActivity.java
@@ -82,11 +82,11 @@ import org.mozilla.focus.widget.FragmentListener;
 import org.mozilla.focus.widget.TabRestoreMonitor;
 import org.mozilla.rocket.component.LaunchIntentDispatcher;
 import org.mozilla.rocket.component.PrivateSessionNotificationService;
-import org.mozilla.rocket.content.BottomBarItemAdapter;
-import org.mozilla.rocket.content.ChromeViewModel;
-import org.mozilla.rocket.content.ChromeViewModel.ScreenCaptureTelemetryData;
 import org.mozilla.rocket.content.HomeFragmentViewState;
-import org.mozilla.rocket.content.MenuViewModel;
+import org.mozilla.rocket.chrome.BottomBarItemAdapter;
+import org.mozilla.rocket.chrome.ChromeViewModel;
+import org.mozilla.rocket.chrome.ChromeViewModel.ScreenCaptureTelemetryData;
+import org.mozilla.rocket.chrome.MenuViewModel;
 import org.mozilla.rocket.content.view.BottomBar;
 import org.mozilla.rocket.download.DownloadIndicatorViewModel;
 import org.mozilla.rocket.extension.LiveDataExtensionKt;
