+++ app/src/main/java/org/mozilla/rocket/chrome/BottomBarViewModelFactory.kt
@@ -1,20 +1,20 @@
-package org.mozilla.rocket.content
+package org.mozilla.rocket.chrome
 
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
 
-class MenuViewModelFactory private constructor() : ViewModelProvider.NewInstanceFactory() {
+class BottomBarViewModelFactory private constructor() : ViewModelProvider.NewInstanceFactory() {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MenuViewModel::class.java)) {
-            return MenuViewModel() as T
+        if (modelClass.isAssignableFrom(BottomBarViewModel::class.java)) {
+            return BottomBarViewModel() as T
         }
         throw IllegalArgumentException("Unknown ViewModel class: " + modelClass.name)
     }
 
     companion object {
         @JvmStatic
-        val instance: MenuViewModelFactory by lazy { MenuViewModelFactory() }
+        val instance: BottomBarViewModelFactory by lazy { BottomBarViewModelFactory() }
     }
 }
