+++ app/src/main/java/org/mozilla/rocket/chrome/MenuViewModel.kt
@@ -1,9 +1,9 @@
-package org.mozilla.rocket.content
+package org.mozilla.rocket.chrome
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import org.mozilla.focus.utils.AppConfigWrapper
-import org.mozilla.rocket.content.BottomBarItemAdapter.ItemData
+import org.mozilla.rocket.chrome.BottomBarItemAdapter.ItemData
 import java.util.Arrays
 
 class MenuViewModel : ViewModel() {
